\name{gen_heatmap}
\alias{gen_heatmap}
\title{Generate a Conservation Heatmap}
\description{
This function is designed to generate a heatmap visualization of residue conservation scores
calculated from a multiple sequence alignment of your protein of interest. The heatmap displays
alignment positions on the x-axis and represents conservation scores
using a color gradient.
}
\usage{
gen_heatmap(scores)
}
\arguments{
  \item{scores}{
    A named numeric vector of residue conservation scores generated by
    \code{conservo()}. The names correspond to the consensus amino acids
    at each alignment position.
  }
}
\value{
A \code{ggplot} object representing the conservation heatmap.
}
\examples{
alignment <- align(
  input = "protein_sequences.fasta",
  method = "ClustalOmega"
)

scores <- conservo(alignment)

heatmap <- gen_heatmap(scores)

heatmap
}