\name{gen_lineplot}
\alias{gen_lineplot}
\title{Generate a Conservation Profile}
\description{
his function is designed to generate a line plot of residue conservation scores calculated from a
multiple sequence alignment of your target protein. The plot displays residue positions on the
x-axis and BLOSUM62 conservation scores on the y-axis, providing a
profile of conservation across the protein sequence.
}
\usage{
gen_lineplot(scores)
}
\arguments{
  \item{scores}{
    A named numeric vector of residue conservation scores generated by
    \code{conservo()}. The names correspond to the consensus amino acids
    at each alignment position.
  }
}
\value{
A \code{ggplot} object representing the residue conservation profile.
}
\examples{
alignment <- align(
  input = "protein_sequences.fasta",
  method = "ClustalOmega"
)

scores <- conservo(alignment)

profile <- gen_lineplot(scores)

profile
}